change outputted maze to use 2D array of rooms & modify user prompt
Printed output of the maze not properly displays whether or not a room has
an open door, as well as sets the default entrance and exit. The user
prompt now checks for valid input and includes the option for the maze
size to be randomly selected. The minimum side length for the maze is
currently set as 4.