create function to generate adjacent rooms and open doors between them
The function getAdjacentRooms takes in the roomNumber of one room and
returns the roomNumber of an adjacent room. The function openDoors
takes in two adjacent rooms and opens the doors between them, which
can be seen in the outputted maze. These fucntions are run once, and
need an implementation to run multiple times until a complete maze is
created.