create algorithm to complete generating the maze
In order to make the maze complete and random, and algorithm is implemented
to check whether adjacent rooms are along a path that connects them before
opening the door between them. Without this algorithm multiple doors could
be opened creating large empty squares in the maze. At this point the
program can generate a complete random maze which can be solved.